fix(packages/sui-mock-provider): fix for app usage separing browser and client
