+++ packages/sui-mock-provider/src/browser.js
@@ -1,0 +1,12 @@
+export {rest} from 'msw'
+
+export const getBrowserMocker = async handlers => {
+  const setup = await import('msw').then(pkg => pkg.setupWorker)
+  const worker = setup(...handlers)
+
+  return {
+    ...worker,
+    listen: worker.start,
+    close: worker.stop
+  }
+}
