--- packages/sui-mock-provider/src/types.d.ts
@@ -1,6 +1,0 @@
-import { RequestHandler as WorkerHandler } from 'msw'
-import { RequestHandler as ServerHandler } from 'msw/node'
-
-type Mocker = SetupWorkerApi & SetupServerApi
-
-export { WorkerHandler, ServerHandler, Mocker }
